sotoxa_backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=3.0.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytesseract>=0.3.8
Pillow>=8.3.2
python-dotenv>=0.19.0